requirements.txt
Django
django_environ
djangorestframework
djangorestframework_simplejwt
drf_nested_routers
freezegun
model_bakery
pyotp
pytest
pillow
django-filter
django-cors-headers
django-debug-toolbar
mysqlclient
pytest-django
redis
celery
djoser
flower
pytest-watch